Move helper function to ioutils
